# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GuardRail API
  description: >-
    GuardRail — Privileged Access Management for browser-based access to web
    management interfaces. All endpoints are under `/api/v1`. Errors use RFC 9457
    `application/problem+json`. Authentication is a Bearer access token (JWT);
    the refresh token is delivered as an HttpOnly cookie.
  version: 1.0.0
  license:
    name: Proprietary
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /search:
    get:
      tags: [Analytics]
      summary: Global search across users, devices, sessions
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, minLength: 2 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResults" }
  /audit:
    get:
      tags: [Analytics]
      summary: Query the audit log (requires log:read)
      parameters:
        - { in: query, name: action, schema: { type: string } }
        - { in: query, name: actor, schema: { type: string } }
        - { in: query, name: result, schema: { type: string, enum: [success, failure, denied] } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/AuditRow" } }
  /reports:
    post:
      tags: [Analytics]
      summary: Generate a CSV report (requires report:read)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [audit, access] }
                format: { type: string, enum: [csv], default: csv }
                from: { type: string }
                to: { type: string }
      responses:
        "200":
          description: CSV attachment
          content:
            text/csv: { schema: { type: string, format: binary } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AuditRow:
      type: object
      properties:
        ts: { type: string, format: date-time }
        actor: { type: string }
        action: { type: string }
        category: { type: string }
        target_type: { type: string }
        target_id: { type: string }
        ip: { type: string }
        result: { type: string }
    SearchResults:
      type: object
      properties:
        users: { type: array, items: { $ref: "#/components/schemas/Hit" } }
        devices: { type: array, items: { $ref: "#/components/schemas/Hit" } }
        sessions: { type: array, items: { $ref: "#/components/schemas/Hit" } }
    Hit:
      type: object
      properties:
        id: { type: string }
        label: { type: string }